Move txversion/txtype outside of cryptonote_basic.h
This is simply enough that it can be used from device (which is above
cryptonote_basic in the deps hierarchy).
